requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
async-tio==1.3.2
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
discord-protos==0.0.2
discord-webhook==1.1.0
discord.py-self==2.0.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.4.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.24.3
openai==0.27.6
OpenAIAuth==0.3.6
packaging==23.1
prompt-toolkit==3.0.38
protobuf==4.22.4
pyfiglet==0.8.post1
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.4
regex==2023.5.5
requests==2.30.0
revChatGPT==5.0.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
soupsieve==2.4.1
spotipy==2.23.0
sympy==1.11.1
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==2.0.2
uwuify==1.2.1
wcwidth==0.2.6
yarl==1.9.2
